Completed Puzzle 1, 2 for stage 1
Completed Basic logic and program implementation, put placeholder for puzzle 2

+++ Assets/Scripts/AutomaticTeleportation.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 using System;
 public class AutomaticTeleportation : MonoBehaviour
 {
-    [SerializeField] int hallwayLength;
+    [SerializeField] float hallwayLength;
     [SerializeField] Transform player;
 
     [SerializeField] Rigidbody2D rb;
